Initial start of uinavigationbar-under-statusbar project
This includes the start of the project for the uinavigationbar that will
be used in our blog post.

This adds a `.gitignore` that should cover the iOS projects so we don't
commit anything we shouldn't.

This is an app without a storyboard so I have removed as much cruft as I
could to give you a good clean place to start. When you run the app you
should see a view that reminds you of UpTech.