Add links to commits for nav bar project
I did this so you can quickly compare the code at the different stages
if you want. This might not work out in the long run if we go back to
project and change things but I'm hoping that it's good enough as is so
this can be useful to someone who is newer.